requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
idna==2.10
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.1
mysql-connector-python==8.0.21
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.1.0
urllib3==1.25.11
